changed create_dir.py - add print to output data directory, list of actions, video_length

+++ data/create_dir.py
@@ -48,62 +48,28 @@ def create_dir(actions, subfolder_numbers, data_dir):
             except:
                 pass
             finally:
-                # w = '{}.iams'.format(data_dir)
-
-                hand_landmark = {0: 'WRIST', 1: 'THUMB_CMC',
-                                 2: 'THUMB_MCP',
-                                 3: 'THUMB_IP',
-                                 4: 'THUMB_TIP',
-                                 5: 'INDEX_FINGER_MCP',
-                                 6: 'INDEX_FINGER_PIP',
-                                 7: 'INDEX_FINGER_DIP',
-                                 8: 'INDEX_FINGER_TIP',
-                                 9: 'MIDDLE_FINGER_MCP',
-                                 10: 'MIDDLE_FINGER_PIP',
-                                 11: 'MIDDLE_FINGER_DIP',
-                                 12: 'MIDDLE_FINGER_TIP',
-                                 13: 'RING_FINGER_MCP',
-                                 14: 'RING_FINGER_PIP',
-                                 15: 'RING_FINGER_DIP',
-                                 16: 'RING_FINGER_TIP',
-                                 17: 'PINKY_MCP',
-                                 18: 'PINKY_PIP',
-                                 19: 'PINKY_DIP',
-                                 20: 'PINKY_TIP'}
-
-                empty_list = []
-                for k, v in hand_landmark.items():
-                    for i in range(3):
-                        empty_list.append(v)
-
-                hand_landmark_x_y_z = [['{}_x'.format(x), '{}_y'.format(y), '{}_y'.format(z)] for x, y, z in
-                                       zip(empty_list[0::3], empty_list[1::3], empty_list[2::3])]
-                hand_landmark_x_y_z = [item for sublist in hand_landmark_x_y_z for item in sublist]
-
                 iams_dict = {'actions': args.actions, 'subfolder_length': args.subfolder_numbers,
                              'video_length': args.video_length, 'Data_Directory': data_dir,
-                             'Data_Subfolder': str(Path("{}/data".format(args.pat))), 'csv_header': hand_landmark_x_y_z}
-
-                # print(len(hand_landmark_x_y_z))
-                # hand_dict = {}
-                # for i in range(63):
-                #     hand_dict[i] = hand_landmark_x_y_z[i]
-
+                             'Data_Subfolder': str(Path("{}/data".format(args.pat)))}
                 w = 'iamsMediapipe.iams'
-                with open(w, 'r+') as f:
-                    f.seek(0)
+                if not os.path.exists(w):
+                    w = os.path.join(w)
+                    f = open(w, 'w')
                     json.dump(iams_dict, f)
-                    f.truncate()
-
+                    f.close()
+                os.path.exists(w)
+                f = open(w, 'r+')
+                f.seek(0)
+                json.dump(iams_dict, f)
+                f.truncate()
                 f.close()
 
-                # w_ = 'csv_header.iams'
-                # with open(w_, 'w') as header:
-                #     json.dump(hand_dict, header)
-                #
-                # header.close()
+    absolute_iams_path = os.path.abspath(data_dir)
 
-    print('Your data Folder has been Created: Folder Name:{}.'.format(data_dir))
+    print('Your data directory is {}'.format(absolute_iams_path))
+    print('Your data is saved in {}'.format(data_dir))
+    print('Defined list of actions are: {}'.format(args.actions))
+    print('Your video will run for {} seconds at every collection'.format(args.video_length))
 
 
 if __name__ == '__main__':
